@@ -1,82 +1,135 @@
 ﻿using System.Text.RegularExpressions;
 using LineCount.Errors;
+using SpanExtensions;
 
 namespace LineCount;
 
 public static partial class LineCount
 {
-    public static async Task<Result<int, DirectoryNotFoundError>> GetLineCount(LineCountData data,  string[] excludeDirectories, string[] excludeFiles)
+    public static async Task<Result<string, DirectoryNotFoundError>> GetLineCount(string path, LineCountData data, string[] excludeDirectories, string[] excludeFiles)
     {
-        var (excludeFileNames, excludeFilePaths, excludeRelativeFilePaths) = ExcludePaths(excludeFiles);
-        var (excludeDirectoryNames, excludeDirectoryPaths, excludeRelativeDirectoryPaths) = ExcludePaths(excludeDirectories);
+        var (excludeFilePaths, excludeRelativeFilePaths) = ExcludePaths(excludeFiles);
+        var (excludeDirectoryPaths, excludeRelativeDirectoryPaths) = ExcludePaths(excludeDirectories);
 
-        if (File.Exists(data.Path))
+        string[] excludeAbsoluteFilePaths = CombinePaths(excludeFilePaths, excludeRelativeFilePaths);
+        string[] excludeAbsoluteDirectoryPaths = CombinePaths(excludeDirectoryPaths, excludeRelativeDirectoryPaths);
+
+        var result = await GetLineCountFromAbsolutePaths(path, data, excludeAbsoluteFilePaths, excludeAbsoluteDirectoryPaths);
+
+        return result.Map(x => $"{x} lines have been found.");
+    }
+
+    private static async Task<Result<int, DirectoryNotFoundError>> GetLineCountFromAbsolutePaths(string path, LineCountData data, string[] excludeFiles, string[] excludeDirectories)
+    {
+        if (File.Exists(path))
         {
-            return await GetSingleLineCount(data);
+            return await GetSingleLineCount(path, data);
         }
 
-        if (!Directory.Exists(data.Path))
+        if (!Directory.Exists(path))
         {
-            return new DirectoryNotFoundError(data.Path);
+            return new DirectoryNotFoundError(path);
         }
-
-        excludeFiles ??= [];
-        excludeDirectories ??= [];
 
         List<Task<int>> filetasks = [];
-        foreach (var file in data.Filter is not null ? Directory.GetFiles(data.Path, data.Filter) : Directory.GetFiles(data.Path))
+        
+        string[] files = (data.Filter is not null ?
+            Directory.GetFiles(path, data.Filter) :
+            Directory.GetFiles(path)).Select(Path.GetFullPath).ToArray();
+         
+        foreach (var file in files)
         {
-            if (!Array.Exists(excludeFileNames, x => x == Path.GetFileName(file))
-                && !Array.Exists(excludeFilePaths, x => x == Path.GetFullPath(file)
-                && !excludeRelativeFilePaths.Any(x => Path.GetFullPath(file).Contains(x))))
+            if (IsExcluded(excludeFiles, file))
             {
-                switch(data.FilterType)
-                {
-                    case CountType.Normal:
-                        filetasks.Add(GetFileLineCount(file));
-                        break;
-                    case CountType.Filtered:
-                        filetasks.Add(GetFilteredFileLineCount(file, data.LineFilter!));
-                        break;
-                    case CountType.FilteredExcept:
-                        filetasks.Add(GetFilteredFileLineCount(file, data.LineFilterNot!, false));
-                        break;
-                    case CountType.FilteredBoth:
-                        filetasks.Add(GetDoublyFilteredFileLineCount(file, data.LineFilter!, data.LineFilterNot!));
-                        break;
-                }
+                continue;
+            }
+
+            switch (data.FilterType)
+            {
+                case CountType.Normal:
+                    filetasks.Add(GetFileLineCount(file));
+                    break;
+                case CountType.Filtered:
+                    filetasks.Add(GetFilteredFileLineCount(file, data.LineFilter!));
+                    break;
+                case CountType.FilteredExcept:
+                    filetasks.Add(GetFilteredFileLineCount(file, data.LineFilterNot!, false));
+                    break;
+                case CountType.FilteredBoth:
+                    filetasks.Add(GetDoublyFilteredFileLineCount(file, data.LineFilter!, data.LineFilterNot!));
+                    break;
             }
         }
+        
+        int rootLineCount = 0;
+        int index = 0;
+        
+        await foreach (var result in Task.WhenEach(filetasks))
+        {
+            if(data.ListFiles && result.Result != 0)
+            {
+                Logger.Log(files[index], result.Result.ToString());
+            }
 
-        var filetaskResults = await Task.WhenAll(filetasks);
-        int rootLineCount = filetaskResults.Sum();
+            rootLineCount += result.Result;
+            index++;
+        }
 
         List<Task<Result<int, DirectoryNotFoundError>>> directorytasks = [];
 
-        foreach (var directory in Directory.GetDirectories(data.Path))
+        foreach (var directory in Directory.GetDirectories(path))
         {
-            if (!Array.Exists(excludeDirectories, x => x == Path.GetFileName(directory)))
+            if (!IsExcluded(excludeDirectories, directory))
             {
-                directorytasks.Add(GetLineCount(data, excludeDirectories, excludeFiles));
+                directorytasks.Add(GetLineCountFromAbsolutePaths(directory, data, excludeDirectories, excludeFiles));
             }
         }
-
+        
         var directorytasksResult = await Task.WhenAll(directorytasks);
         int directoriescount = directorytasksResult.Where(x => x.IsSuccess).Sum(x => x.Value);
 
         return rootLineCount + directoriescount;
     }
 
-    private static async Task<Result<int, DirectoryNotFoundError>> GetSingleLineCount(LineCountData data)
+    private static string[] CombinePaths(string[] excludeFilePaths, string[] excludeRelativeFilePaths)
     {
-        return await (data.FilterType switch
+        string[] excludeAbsoluteFilePaths = new string[excludeFilePaths.Length + excludeRelativeFilePaths.Length];
+
+        ConvertRelativePathsToAbsolute(excludeAbsoluteFilePaths, excludeRelativeFilePaths);
+
+        Array.Copy(excludeFilePaths, 0, excludeAbsoluteFilePaths, excludeRelativeFilePaths.Length, excludeFilePaths.Length);
+        return excludeAbsoluteFilePaths;
+    }
+
+    static void ConvertRelativePathsToAbsolute(string[] absolutePaths, string[] relativePaths)
+    {
+        ReadOnlySpan<char> workingDir = Environment.CurrentDirectory;
+
+        if (workingDir[^1] == Path.DirectorySeparatorChar)
         {
-            CountType.Normal => GetFileLineCount(data.Path),
-            CountType.Filtered => GetFilteredFileLineCount(data.Path, data.LineFilter!),
-            CountType.FilteredExcept => GetFilteredFileLineCount(data.Path, data.LineFilterNot!, false),
-            CountType.FilteredBoth => GetDoublyFilteredFileLineCount(data.Path, data.LineFilter!, data.LineFilterNot!),
+            workingDir = workingDir[..^1];
+        }
+
+        for (int i = 0; i < relativePaths.Length; i++)
+        {
+            absolutePaths[i] = $"{workingDir}{Path.DirectorySeparatorChar}{relativePaths[i]}";
+        }
+    }
+    static bool IsExcluded(string[] excludeFilePaths, string file)
+    {
+        return Array.Exists(excludeFilePaths, x => x == Path.GetFullPath(file));
+    }
+
+    static Task<int> GetSingleLineCount(string path, LineCountData data)
+    {
+        return data.FilterType switch
+        {
+            CountType.Normal => GetFileLineCount(path),
+            CountType.Filtered => GetFilteredFileLineCount(path, data.LineFilter!),
+            CountType.FilteredExcept => GetFilteredFileLineCount(path, data.LineFilterNot!, false),
+            CountType.FilteredBoth => GetDoublyFilteredFileLineCount(path, data.LineFilter!, data.LineFilterNot!),
             _ => throw new NotImplementedException(),
-        });
+        };
     }
 
     public static async Task<int> GetFilteredFileLineCount(string path, Regex regex, bool filterResult = true)
@@ -149,20 +202,13 @@ public static partial class LineCount
         return count;
     }
 
-    static PathPatterns ExcludePaths(string[] excludeFiles)
+    static PathPatterns ExcludePaths(IEnumerable<string> excludeFiles)
     {
-        List<string> excludeFileNames = new List<string>();
-        List<string> excludeFilePaths = new List<string>();
-        List<string> excludeRelativeFilePaths = new List<string>();
+        List<string> excludeFilePaths = [];
+        List<string> excludeRelativeFilePaths = [];
 
         foreach (string filename in excludeFiles)
         {
-            if (!filename.Contains(Path.DirectorySeparatorChar) && !filename.Contains(Path.AltDirectorySeparatorChar))
-            {
-                excludeFileNames.Add(filename);
-                continue;
-            }
-
             if (Path.IsPathFullyQualified(filename))
             {
                 excludeFilePaths.Add(filename);
@@ -172,8 +218,11 @@ public static partial class LineCount
             if (filename.StartsWith("./"))
             {
                 excludeRelativeFilePaths.Add(filename[2..]);
+                continue;
             }
+
+            excludeRelativeFilePaths.Add(filename);
         }
-        return (excludeFileNames.ToArray(), excludeFilePaths.ToArray(), excludeRelativeFilePaths.ToArray());
+        return (excludeFilePaths.ToArray(), excludeRelativeFilePaths.ToArray());
     }
 }
